breaking classes into files ala PSR4
